renamed modules to lowercase
